Update mcmc.py to allow for a prior on logU
Allow for a prior on logU (i.e., n_H, the total hydrogen number density).  The current constraint is a Gaussian prior for logU with (approx) mean = -3.0 and sigma = 0.5 (values based on Lehner et al. 2013; Wotta et al. 2016; and Wotta et al. 2017).  It is then converted into a prior on n_H by integrating the UVB at the absorber's redshift (currently only two options for the UVB: HM05 and HM12 because those are the two UVB's I have).  Yes, it's currently hard-coded.  This may change.

You do not have to use this option (its default is "False").  But it's VERY useful when you can't constrain the metallicities using the ions as they are (e.g., all upper limits, leading to an upper limit on the metallicity).