Now allows for "no optimization" with --wotta
Previously, the --wotta option implied the args.optim='guess'.  Now, if the initial guess (in "MCMC_initial_guesses-run_me.dat") for either met or dens are "False", then args.optim=False.  This allows for the first, quick, full-grid search for a global maximum, which can then be followed up with a better guess.